M2-6040: Move the new test functions back to axois-mocks.ts
Putting them in the testUtils.ts file caused `@jest/globals` to be included in the bundle sent to the front-end, which in turn breaks the app.

In the future we should completely separate test functions from those being used in app code
